requirements.txt
discord.py>=2.3.0
praw>=7.7.0
groq>=0.4.0
python-dotenv>=1.0.0
